% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dcStockR.R
\docType{package}
\name{dcStockR-package}
\alias{dcStockR}
\alias{dcStockR-package}
\title{Dimensional Charting of Stock in R}
\description{
Dimensional Charting of Stock in R
}
\seealso{
\url{http://dc-js.github.io/dc.js/}
}

